Fix bug in subtree tool that caused first subdir to be stripped
The internal file representation for the satdead branch missed the ./
prefix that is part of the anotated files used in the old reference
checker. This problem confused the tree drawing tool so it stripped of
the first component of the path. Now we add that prefix again
correcting the output