undertaker: change behaviour of model loading
Now, the default is, that no models are loaded. The user can specify a
model directory with -m, then all .model files within this directory
are loaded. If the Argument to -m is a file, it is loaded as a single
model file.

With -M the user can specify the main model, which is used as primary
model.

Replace options as following:

-s: "" (skip models)
-a: "-m model_file -M model" (use just a single model)