Tune coverage configuration calcuation
For coverage analysis, items in the missing set cannot be
enabled. Therefore, the formula used for finding the satisfying
selection of variables should not consider them as 'free' variables, but
set them to "not selected". This is similar to finding defects, where we
do the same.

This should heavily reduce the set of calculated configurations further!