undertaker-calc-coverage: automatically rewrite .pi -> .c in worklists
The .pi files are partially preprocessed files produced as debug output
from typechef. As they are rather large (>5GB data!) our experiment is
run in two stages: in the first run, we have collected all
configurations in various modes and stored them in cpio archives. In the
next step, we extract these configs from the archives in a clean working
copy. With these commits, we don't really require the .pi files in the
working copy. Moreover, we want to do the actual analyis on the real
files anyway.