SatChecker: implement BaseExpressionSatChecker
This SAT Checker uses the same formula, parses it only once and can be
used to query the same clause set multiple times. This improves the
performance tremendously. It fills Picosat with the formula once and
uses the feature of picosat to work on the same working set of clauses
multiple times.