test-ConditionalBlock: Do not initialize CppFile globally
This patch ensures that the constructor of CppFile is run after
execution has passed to main(). The reason is that CppFile uses PUMA in
its constructor to parse the given filename. However, Puma itself uses
public constructors to initialize its StrHashMap(). In the case at hand,
Puma is used to parse a file before the HashMap is executed, which leads
to a segmentation fault.

Change-Id: Iff14e4c36abaedefe385a6c719e89c0eca46d143