undertaker: Add an API for querying the type of a configuration item
Unfortunately, the type of an item is not stored in the models as
produced by rsf2model, but (currently) is only available in the rsf
files. While extending the model would be an option, this change leaves
rsf2model alone. Instead the undertaker will now parse the .rsf file in
addition to the model file.

We therefore no longer remove them in undertaker-kconfigdump, and
make sure to safe them next to the .model file with an .rsf suffix.

Change-Id: Ieca73e6dcde8391e92084dad153484144cfe187f