coverage: ignore non-boolean items in partial configurations in kconfig mode
If we encounter an item that is neither tristate nor boolean, we cannot
set it to some value and expect kconfig to do something useful with
this. In fact, current kconfig will now even emit a warning message for
that.  In kconfig output mode, we therefore rather skip these
items. Hopefully, this further reduces expansion errors.

Change-Id: I60fac1e79ec2b25ad9eb33beafb7da258cdd3a2e