undertaker: exec output mode with commented out blocks
In the exec output mode of configuration formatter of the SatChecker
class the calculated coverages are piped into a process. This change
changes the output to comment out the code blocks instead of omitting
them in order to preserve the linenumbers. This will be useful because
for using the coverage approach together with coccinelle.

This feature can't be tested with out test-suite properly, because it
would result in a recursive quine problem.

Change-Id: I8f15c20d93953ab464eea833badcb10b05b851e3