undertaker: new Logging framework
To unify the ad-hoc logging a new Logging framework is introduced,
which can be used as a standard c++ stream. It knows 4 predefined log
levels (debug, info, warn, error).

All nesseccary places should have been updated. In the testcases only the (now)
unnesseccary -q flag was removed.

Additionally, the semeantics of the commandline option -v is changed; it
now increases the verbosity level. In order to identify the undertaker
version, the option '-V' is provided.

Change-Id: I552583c855788e8d04c0c1d8f73fdd4eaa202c18