Build-System: Avoid concurrency in top-level Makefile
This project uses recursive make invocations, mainly because it is an
aggregation of several subprojects. In order to avoid concurrency
issues, like compiling the same targets several times and taking only
the last result, we avoid concurrency in the top-level Makefile by
adding the gnu-specific .NOTPARALLEL target.

Parallel building in sub-level makefiles is still maintained,
though. This change only prevents targets for different tools to be
executed independently.

Change-Id: I4bde1a46a3ddf8d58525f9274472d1a89a1e3ca9