undertaker: add new working mode: -j cppsym
Introduce a new working mode called "cppsym" to print out all CPP
symbols that are used in the #ifdef expressions of the given source code.

Optionally, a model file can be given.  In this case additional
information such as the type of a symbol (e.g., BOOLEAN or TRISTATE) and
its presence in the configuration model (i.e., MISSING) is printed.

Change-Id: I51315345103324020c4ca11b8f2db2864391d20f