golem: factorize error checking of commands and calling Linux makefiles
This change reduces code duplication of error handling, making the code
easier to read and write. Moreover, the fairly complicated logic behind
calling make properly with respect to configuring for foreign
architectures is now properly hidden in a new method call_linux_makefile()

The existing tools 'undertaker-calc-coverage' and parts of 'vampyr' are
updated to use these new infrastructure.

For testing the new methods call_linux_makefile() and
guess_arch_from_filename() parts of the newer unittest package of
python2.7 are used. In order to remain backwards compatible with python
2.6, the added test requires the python-unittest2 package installed.

Change-Id: I80c140b1480dfbd9c403a5a61d3f21e487e073f8