undertaker: enhance cppsym mode to do csv-like statistics
With this change, undertaker counts references to and redefinitions, of
CPP macros in source files. Moreover, the macros are normalized so that
the resulting statistics actually count references to features in
Kconfig.

If a model is given, the statistics also include the type as stated in
the model.

The output format is CSV-like so that it can be easily parsed.

Change-Id: I51673489df71ef07870a0cedab3983f6e087c9e0