golem: implement linux makefile feature inference mechanisms
golem is extended to infer implications (i.e., constraints) for compiled
source files from Linux makefiles. The constraints indicate what
features can be assumed to be selected if a given source file is compiled.

The basic approach is to collect what files are compiled with different
feature selections, and deduce the implications from the resulting
intersections of the filesets. When a directory is enabled by a feature,
the process is done recursively.

The undertaker parser and logic engine is extended to take these
constraints into account.

This extraction is integrated into undertaker-scan-head.

Change-Id: Ib1b46528cbe21b019a1cf0c4070ecd89cc82a3fa