rsf2model: choice implications are always true
Since the implications (like the cardinality of the items in the
choice) are always true, they have to appear in the ALWAYS_ON list.

For boolean choices this is quite straight forward, just include it in
the ALWAYS_ON list.

For tristate choices, we introduce a new variable
CONFIG_CHOICE_%d_META. This meta variable is added to ALWAYS_ON and
the CHOICE_%d and CHOICE_%d_MODULE depend on this meta variable.

Reported-By: Christian Kästner <christian.kaestner@uni-marburg.de>
Change-Id: I2ea125ee9e99c6480577ba677ac893e9c618bff0