undertaker: Produce a valid configuration also for files without variability
The point of the coverage mode is to produce a set of configurations
that maximize the configuration coverage. This can also be calculated
for the trivial case: a file without #ifdef blocks requires exactly one
configuration for maximal configuration coverage.

This commit is a bit hack-ish: It modifies the CPPFile structure to
consider the top_block "B00" as regular block.

Change-Id: I23ad747b742e429a843b92f70837d8cc720a21f0