golem: fix Makefile.list and golem -i
Because including the Linux top-level makefile can be influenced by
the active loaded configuration, the auto.conf must be included
beforehand and inclusion in the top-level makefile must be prohibited.

Additionally $(lib-y) should also be used in printing and doing the
recursion.

golem -d output is extended in debug mode to print the contents of
the check_subdir function.

Change-Id: Ib0e12b6170173df5a68ae8a507bbda7fd8d3427c