vampyr/-ae: also calculate the dead and undead blocks on the go
In order for this to work properly, we need to be more careful when
calculating conditional blocks. This change extends the .coverage
reports with the field 'configuration_blocks', which now contains only
blocks that are configuration controlled (i.e., contain a reference to a
CPP that starts with 'CONFIG_'). The original field 'total_blocks' now
contains both types of blocks.

The testcase ensures the proper counting of blocks in zizler and
undertaker. The basic idea is to check that the highest block number
found by undertaker and zizler is the same.

Change-Id: I57893402ae1a09796d5e3c9c260f57a84f8b116a