python: refactor calculate_worklist out of golem
calculate_worklist calculates a sanitizes worklist from a list of given
arguments. This method is intended to be passed lists from the command
line. It is passed a list of filenames, which are tested individually
for existance. Non-existing filenames are skipped with a warning.

Optionally, the given files can be interpreted as worklist. In this
case, each line in the given files is tested for existance and
considered for the worklist.

In any case, the caller can rely on a set of filenames which are
good for processing.

Change-Id: I27cfce2f40a83f59b46fb45f1ce9bba511f1f7fb