vampyr: relocate variability statistics and analyses from golem
This relocation allows the analysis mode to be applied with the bare
framework as well. The analysis is vastly improved compared to golem:
The file .coverage.$ARCH gives a more detailed overview what blocks
where enabled during the coverage analysis. Also the coverage of
allyesconfig is calculated, both in terms of block count as in terms of
line count.

This change allows more detailed output during analysis. In particular,
the file .coverage.$ARCH details what blocks where enabled during the
coverage analysis. Also the coverage of allyesconfig is calculated, both
in terms of block count as in terms of line count.

The option -a now always expands all configurations. A later change will
allow checking a specific stdconfig.

Change-Id: If833dee604cc4f3394c2570a25546a3858261d92