fakecc: a compile-to-/dev/null c compiler
Introduce a gcc commandline-compatible tool. It is intended to be used
with kbuild to allow source files to be checked without calling a real
compiler. This allows to analyze source files sith static checkers, such
as sparse, from the Linux HAL for architectures for which
no (cross-)compiler is installed on the system.

This change also integrates fakecc in the sparse checker of vampyr.

Change-Id: I32356c74c15c66816b47d2980c67d33a4e46cbb0