golem: Fix makefile dependency extraction on unconfigured trees
Golem's FileSymbolInferencer wrongly assumed that apply_configuration
would work on unconfigured trees. By configuring 'allnoconfig', a
standard linux target that all architectures implement, less assumptions
are being made on the state of the Linux tree is made.

This bug was found by the change in commit f403f42f that causes
apply_configuration to fail if '.config' is not present. In order to
make the code more resilant here, the standard configuration
'allyesconfig' is set in this case.

Change-Id: I4a545aaf5c5046030b334e991f9d2b96d6ef50da