ziz: copy test patterns and re-enable ziz testsuite
The testpatterns obviously were out of date and do not match the actual
code. Obviously, we would need to review them and possibly fix the code
if some of them do not make sense. However, as ziz is largely
deprecated, this is probably not worth the effort.

We are interested in those test patterns for two reasons: first, the
could serve as templates and inspiration for predator (which in the end,
could replace zizler one day). Second, we are interested in pattern
changes as they help to understand code changes in ziz.

Change-Id: Iffeeb2c1455a7899b107287db239f0ca81d32412