vampyr: Refactorings in KbuildFramework and Configuration subclasses
This change cleans up the parameter proliferation in
KbuildFramework. Instead, the parameters arch, subarch,
expansion_strategy and coverage_strategy are passed via the options
dictonary.

Also, implement a filename() method for the Configuration subclass to
retrieve its corresonding filename. This requires additional refactoring
in the __init__ classes of the subclasses.

Change-Id: I1ddcb69c0f0cdb7ad51e43e915dbaa1df277a1f2