undertaker: fix counting of blocks in the SatChecker class
While commit e1e7f90addb15257520937c7782710caf56d4101 fixed the
configuration for files without variability, it also broke the
presentation of the results. With the Block "B00" inserted as first
block, the block "B0" (and its subsequent blocks) no longer start at
index 0 in the bit vector but are shifted by +1.

Technically, the previous implementation mapped both B00 and B0 to the
same index. Therefore the block "B00" is never counted if "B0" is
disabled, which means that full coverage was never presented. Luckily,
this only had influence on the output of stdout, and not calculated
results.

Change-Id: Idd0d698821a3c5432a5818edd57378ac05ece592