undertaker: coverage white- and blacklist support
This change extends the coverage algorithms to consider white- and
blacklists. All whitelisted features are set to 'y' and all blacklisted
to 'n'. The lists are read from simple text files (one item per line)
that can be set with -W (whitelist) and -B (blacklist) command-line
options.  Blacklisted features are now listed in the model as
"ALWAYS_OFF".

Change-Id: Id7d849c85999eaf5221f6f7bd1808116d4383b21