undertaker: rename whitelist option to ignorelist
Some background: Previously, in the Eurosys paper, we needed a list of
items that should not be reported as defects. Or more clear, the
affected blocks must not be reported as defect. This is what the failing
testcase tests. The implementation idea was to make those items 'free
items' in the sat checker.  In the end, this means that the current
"whitelist" option does actually specify an "ignorelist", a list of
variables that will always end up as free, unconstrained variable.

This change extens the KconfigWhitelist class to hold three lists (each
of them as a seperate instance of the class): An ignorelist, a whitelist
and a blacklist. The former whitelist functionality (option -w) is now
superseeded by the option -i (as ignorelist). The lists -W (for
whitelist) and -B (for blacklist) currently have no function, the serve
as stubs for upcoming changes.

Change-Id: Ie27bad6e855d536ac502687967e07dc954d42ed1