golem/inferences: rewrite the inference mechanism, port to fiasco
In order to support parallelization and other build systems than
Kbuild, the inference mechanism was completly rewritten. The needed
primitives, which are build-system dependent, are seperated to
inference_atoms.py. These primitives are used by the Inferencer
class (inference.py) which uses the same algorithm for probing, but
does in addition spawn multiple threads for probing.

Change-Id: I3b633f7e9cbd0246137b01205b4ab9e0c9e3edf1