vampyr: Always count both kinds of #ifdef blocks
This change makes the variability analyzer always count
configuration-dependent and non-configuration-dependent blocks in the
result dict. Also, a number documentation strings are updated and a
testcase is added. While writing the testcase, a number of minor bugs
have been fixed.

This changes the output of the .coverage files slightly, so experiments
need to be adapted!

Also note that experimentes need to take special care
wrt. configration-dependent blocks if only those shall be analayzed!

Change-Id: Ib72ad88821eb4cd4efe934e15afefefe0294ba9a