vampyr: rework configuration driver tool evaluation
With option --call, vampyr drives the specified tools in a configuration
aware way in order to find bugs in code that is only conditionally
compiled. In order to evaluate the effectiveness of our approach, this
change improves the message evaluation so that only "interesting"
problems are shown.

For instance in Linux, messages that are not shown by the reference
configuration (e.g., allyesconfig) are most relevant for us. Therefore,
the presentation of the found results filters away messages that are
also present in the reference configuration that is specified with the
--stdconfig switch.

Change-Id: I9bea682f0ab9ac060a561da14c6b43a0267d87c3