vampyr: rewrite message evaluation logic
Configuration dependent bugs are bugs that occur only when a specific
Kconfig configuration has been enabled. The vampyr tool enables
developers to find such bugs. First, it uses the undertaker tool to
calculate a set of configurations for a given file and applies a
traditional static checker on each configuration. If the static checker
reports an issue in some but not all configurations, then vampyr reports
it as configuration dependent defect.

The challenge is that for many files, the undertaker finds exactly one
configuration. This happens for instance in files that do not use #ifdef
statements at all. Note that it is still worth to investigate files
without #ifdef blocks, because the 'allyesconfig' configuration can
still have an effect on header files that are directly or indirectly
referenced via the #include CPP statement. Therefore, we do add the
allyesconfig configuration additionally, because it might lead to a
different compilation unit than the artificially created "undertaker"
configuration.

The other challenge is cocci/spatch. Coccinelle always scans all
files, also the ones not selected by allyesconfig. In our approach, we
prepare the source files in a way that all not selected configurations
get commented-out in order to disable cocinelle s heuristics wrt. #ifdef
blocks and force the tool to parse exactly the configuration the
undertaker has calculated. For this, we need a different strategy: We
need to filter out all defects that have also been found with the
unpreprocessed source code. Technically, this is achieved by selecting
allyesconfig and run the scan (with vampyr-spatch-wrapper), even if the
file is not built by allyesconfig.

This means that we need tool dependent strategies. With Gcc, we add the
default configuration only if that configuration does actually build
it. If the file is not built by the selected configuration, then we are
likely to see compilation errors anyways.

With cocci, we always add the default configuration in order to always
have a vampyr_spatch_wrapper run on the unprocessed source code. First,
cocci does not abort if the file doesn't compile and second, this
ensures that the baseline messages appear in the set.

This patch also changes the semantics of vampyr's --stdconfig switch: It
now defaults to 'allyesconfig' so that the is always checked additionally.

Change-Id: I5d28a8155dc9290c7accf10d73d80bd7175a6211