undertaker: Fix a memleak in the SatChecker class
After building a boolean expression with BoolExp::parseString(), it must
be freed with the delete operator as soon as it is no longer
needed. Reusing it for later steps is also dangerous, as CNFBuilder may
change the internal state of an expression. Therefore, special care must
be taken when reusing a parsed BoolExp.

Change-Id: I5576a4141bb498e9d1452337984af3b60eec8d37