undertaker: bugfix in partial configuration output
When creating the partial configurations in the 'coverage' mode, proper
care is required with items that are not mentioned in the model. Such
"missing" items (i.e., they are not mentioned in the model) must not
appear in the created '.config' files. This patch now makes
SatChecker::formatKconfig() use the correct missingSet, which is filled
at the time of model slicing in the coverage analyzers instead of an
freshly created, always empty set.

Change-Id: I9091e7c0d097a01abb4c4c77cae4e86d5632cf5f