rsf2cnf: fix problems by joining cnfs.
If a variable was declared but never used in the model (happens
for example if the variable is whitelisted) and its cnf-id was higher
the id of the highest used variable, maxcount wasn't icremented.
This causes no problems in most circustances, since these vars are
free. But if an CNF Builder will attach clauses later, this variables
could be remapped.

Change-Id: Ide50c93ca5513dbd422b7a636058fcfbc21687e3