undertaker: Tune detection of dead and undead blocks
The expression of an #ifdef blocks needs to contain at least one item
that is in the configuration space of the model.  Otherwise, the block
is not configuration dependent and must not be reported as configuration
defect.

This behavior makes sense when cross-checking across different models,
and is thus enabled by default if several models are loaded, and
disabled otherwise. With the new switch '-s', this new behavior can be
forcefully enabled.

Change-Id: Idac8c652bc63062e3b59d64ada5b488295aa4b0a