picosat: Introduce picosat version 936
This change introduces a copy of the picosat sources and makes
undertaker link unconditionally against that copy. This is because the
latest version of picosat introduces significant incompatible API
changes. Therefore, we ship a version of picosat that we have tested
extensively for the convinience of our users.

We will update picosat to a later version in an upcoming release.

Change-Id: I9aa4bdf0f88b8b5f4e6cf8b7bfec9f4ae5c3e126