vampyr: handle existing partial configurations with option '-k'
This change enables vampyr to reuse partical configuration that have
been calculated manually. Without the option -k, vampyr will remove
any pre-existing partical configuration that may be in the source
tree.

In particular, this is required for scenarios, in which the partial
configuration needs to be calculated on a modified source tree, and the
actual analysis is to be done in the real source tree.

Change-Id: Ic64a72d890c8a1f8945d0feb4e59169744ae253e