undertaker: always include formula in defect reports
In some cases, the defect reports miss the dependencies found by file
inferences and missing constraints. This change causes that always the
complete formula makes it to the report. In some cases, this may
considerably enlarge the defect reports.
Additionally, empty lines at the end of the defect reports are now
omitted.

Change-Id: I9632b054b857e453d0fbe485a1edc76891c838e6