tailor: optional parameter to provide strippable path
So far, undertaker-tailor always tries to derive the prefix which
needs to be removed from the absolute path given by the debug
information to get a path relative to a kernel source tree on its own.
On Ubuntu e.g. this strippable path might be "/build/buildd/linux-3.2.0"

However, some of that path information could be in a non-canonical
form. We found that on Fedora one particular path was something like
"/foo/bar/////boo/kernel-tree", which (with our current heuristics)
leads to only stripping away "/foo/bar/" and thus breaking the
analysis of the undertaker tool.

This change adds
a) canonicalization for paths from debug information
b) the option to provide the part which needs to be removed on the
command line

Change-Id: I9361dbe8e9a4d1730a5af1924f702798d7320d94