undertaker-kconfigdump: list arches with suffixed Kconfig-Files
An architecture ('um') has just Kconfig-Files with suffixes like
.common/.net/.debug but no Kconfig-File without suffix.
Our current way to detect architectures didn't include 'um' and was
therefore added manually to the architecture list. The 'new way' is
just a minor modification and does now include 'um' without manual
addition to the list and is therefore more of cosmetical nature.

Change-Id: I2af4bfede3c0bf5244ef8c60d387e9e3ac936c64