golem: fix identification of version for ARCH tile
When calling golem while tile architecture is set in the ARCH
environment variable the identification of the project version fails
when no '.git' directory is present, hence 'git describe' is not
possible.
This happens because get_busybox_version is called first and this
function tries to identify the version through a make call after
recognizing no '.git' is present.
Tile itself has the property of cancelling make when the envirionment
variables TILERA_ROOT or CROSS_COMPILE are not set.
get_linux_version normalizes ARCH and SUBARCH to i386 thus calling
make to identify the project does not fail when tile is set
previously.
Thus, trying to retrieve the linux-version first fixes this problem.

Change-Id: Ibcdeff1abbc75cb537b298e3c45dee1395b4affd