Undertaker: introduce no_kconfig defect distinction
Ifdef-blocks can be related to Kconfig items if symbols are containing
the prefix 'CONFIG_'. Others are therefore not related to Kconfig
items. These two different kinds of items were not distinguished until
now.

This change introduces the 'no_kconfig' defect, indicating
that the defect is not related to a Kconfig item. It allows a
more clear differentiation between defects. New testcases
are added, others are updated.

This is a rework of change 1489
(Commit: 98aad7c4659382dee7d621e80084fa5914495e83).

Change-Id: Ia01777759837512ebca1b6c772843365bc4090b9