Undertaker: fix no_kconfig defect detection
(1) Some cases resulted in inconsistencies of reports so that
'kconfig.no_kconfig' reports were possible. These issues are
addressed by introducing 'NoKconfig' into the ENUM in the analyzer.

(2) no_kconfig defects are as a matter of fact global defects.
Consequently (and as a result of (1)), these defects are now
reported as 'no_kconfig.globally'. The nature of this defect class
should now be clear to every user.

(3) The 'no_kconfig.c' testcase has been rewritten to test all
necessary basic constructs. Other affected testcases are
updated.

Change-Id: Idb5699814e97af4306e936f9c7517009662550ca