undertaker-kconfigdump: fix exclusion of arches in conditions
While creating the models with undertaker-kconfigdump, we had the
intention to replace the 'CONFIG_$ARCH'-Entry with an entry containing
conditions to exclude all other architectures (e.g. CONFIG_X86
"!CONFIG_ALPHA && !CONFIG_ARM && [...]").

There was a misconception while creating the code to remove the
'CONFIG_$ARCH'-Entry. It was assumed that there were no conditions
after this entry. Thus, a line was matched which only contained the
entry (e.g. 'CONFIG_X86\n') and the line was not removed if there were
conditions but the entry containing the exclusions was added
nevertheless.
Therefore we had two entries for almost all architectures (except
powerpc, um, sh) on which only the first entry in the file was used
for our calculations.

With this change, these two lines are merged.

Change-Id: Ia05492715f64102128cec311bdce1e2371e5f4c8