undertaker: remove '-d' option/simplify decision coverage application
With the '-d' option it was possible to enable decision mode
preprocessing for other modes than coverage analysis. We introduced
this option for consistency reasons but, nevertheless, it doesn't make
much sense to analyse with decision preprocessing for most of the (-j)
working modes. Therefore we remove this option and add a new workmode
'cpppc_decision' to provide information about the block precondition
structure if doing decision coverage analysis.

Additionally, the way how decision preprocessing is applied is
simplified.

Change-Id: I56741279b73e28f20a5254dba7c4fc05046a2dc3