undertaker: remove unused list member within BlockDefectAnalyzer
The _OKList member, which stores on which architectures the Block is okay (=not
dead), within BlockDefectAnalyzer is unused, thus we can safely remove
it. It's most likely an artifact of debugging.

To be able to easily revert this change, if it's needed once again,
i've put it's content into a seperate commit.

Change-Id: I7c7c23d85c34ce07eed6c81dd9a44109fec7bc92