model.py: is_defined method for RSFModel
The method indicates if the feature is defined in the RSFModel. It
is somehow a basic check if a string is in the model's keys, but I
consider it to be more readable in the code.

Additionally I fixed a small issue in the parse-method and followed
some pylint warnings.