undertaker: add workaround for ModelContainer to speedup batch-mode
With the c++11 change, it was ensured that the ModelContainer and all
the models it is containing, are properly destroyed when the program
is successfully terminating.
However, since we are (currently) parallelizing with a 'fork' for
every file we want to analyze, the destruction of the ModelContainer
happens for every forked process (seperate adress spaces).
While this is the correct behaviour to free all allocated ressources,
this is sadly slowing down the destruction of the child-processes on a
massive scale.
This is slowing down a cppsym-analysis for a huge worklist (>22k) from
about 20 minutes to over 3 hours.
While it previously took around 750ms to do the cppsym-analysis and
then destroy the child process (including the destruction of the
ModelContainer), it only takes around 30ms to do the same without
destroying the ModelContainer and leave the cleanup for the OS.

Until we parallelize on a lightweight thread level, we have to
workaround this problem.

Reported by Bernhard Heinloth.

Change-Id: I00f47d68c6a54137c66a704121a1fbe37b4f9ffd