undertaker: refactor BlockDefectAnalyzer
The static methods are moved into a BlockDefectAnalyzer-Namespace and
a abstract base class 'BlockDefect' is created.
This base class is implementing the mutual behaviour of the
DeadBlockDefect / UndeadBlockDefect classes and minimizes virtual
function calls.
Additionally, the UndeadBlockDefect class is now inheriting from
BlockDefect instead of DeadBlockDefect, which leads to a clean
seperation of concerns.

Change-Id: I986b767d7a4180a2b18a1092b51c300baea718b3