undertaker: remove abstract CNF base-class
After several refactoring passes (c++11 and others), the PicosatCNF.h
header provides the same methods as the CNF.h without exposing
anything else.
Therefore there is no need to use an abstract base-class as interface.

Additionally, the USE_ZCONF defines are gone now and the related
functionality is implemented in the only place it is used, the
SymbolTranslator.cpp.

Change-Id: Iae7acf0ffcd40de3e69502837f9377885d4cc5e9