undertaker/golem: normalize inference symbols in golem
We normalized the filename of the analyzed files in several places
within our code. This made our code not just harder to maintain but
also inconsistent:
The CNFBuilder (to be more precise, the BoolExpParser) has problems
with several chars (like: ":+/,-") in symbols within formulae and
fails.
When we transform the inferences for CNF-Models in rsf2cnf, we had to
normalize the inference symbols (which identify the filename to which
the formula belongs) and store that normalized symbol in the model.
However, in RSF based models, we just append the inference formulae to
the model, which leaves us with two different ways of storing
inferences in the models and more complicated code.

To fix this, "golem -i" now prints the normalized filename (the path
to the file is still recognizable) and undertaker normalizes the
filename _once_ when the normalized file-Variable is accessed through
a getter method from the CppFile-Object.

Change-Id: I52cff842fc8cd1fbb42543713d912c5111c5b38e